requirements.txt
openai
python-dotenv
fastapi
uvicorn
requests
beautifulsoup4
